backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
qdrant-client==1.11.0
sentence-transformers==3.1.1
google-generativeai==0.8.3
beautifulsoup4==4.12.3
requests==2.32.3
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
httpx==0.27.2
